Revert "Use proper API and switch/case for bta search/discovery ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/58771d58-0f69-4d77-be20-70159cdd4c81). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6fb4a8b09f6049358ea3c18c74c01bc64d8c4e3b