Disable aggressive optimizations for Rust host binaries.
The performance of host binaries doesn't really matter, so stop building
them with aggressive optimizations such as LTO and low codegen-units,
as this slows down the build unnecessarily.

On my machine this reduces build time for host-bionic crosvm (required
after every bionic change on Cuttlefish) from >2m to ~15s.

It turned out that libbt_shim_ffi was using LTO to cover up the fact that
it was missing some symbol dependencies that ended up being unused in the
final binary. In some cases, the missing dependency would be difficult
to resolve, e.g. external/rust/crates/nix/src/sys/memfd.rs depends on
memfd_create which is missing from our ancient copy of glibc. So let's
keep that working by forcing LTO on for that library.

Change-Id: I667cddd09f9c3b5b16372aaef542177f5c46511a