Revert "[RootCanal] Enable LL_PRIVACY"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/932eeabd-3691-4337-8623-d801ba1644c2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I17b278530a8c3a23f3098fbc21055341a3ff3558