Revert "[RootCanal] Enable LL_PRIVACY"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6ed30b33-0bb6-442c-a790-0b229678932e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I09166113f806ec1d4448d7493dbd3aa407e38255