Revert "[RootCanal] Enable LL_PRIVACY"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a57b3c8c-27f5-4205-885a-3e6b919f3420). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id70fef7d27a02ed153f5526d67e164c8fee36f06