Fix Map test with wrong dependency to hidden api
Context:
The BluetoothInstrumentationTests is not setup with correct rules.
It should not use reference to jarjar name and instead should jarjar the
code directly. To do so we need to change the following:
* Link against a non jarjared bluetooth app
* use fwk-bluetooth-pre-jarjar instead of the .impl

This CL:
FastXmlSerializer is no longer used in Bluetooth since aosp/1799788 (see
it for justification on why). We now need to remove the dependency from
test as this will not build once we start using jarjaring (or we will
need to add another static dependency to the test)

Test: atest BluetoothInstrumentationTests
Bug: 223837004
Change-Id: I71ead70dff14a5916c5fe64c67c7ad99697b50ef