Revert "RootCanal: comparaison -> comparison"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8cac2fbf-36a6-4f47-b40e-7aad4c92906f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I354ef6425db4626f2ac8dbf607e71242077280ec