Revert "RootCanal: comparaison -> comparison"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/95f8ad1f-d932-4e30-8f76-fbdc58b84723). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I162338fc0ff156e5484a1da0d1f749a3cac5907b