LeAudio: Fix possible volume control delay
Currently some volume control operations are not confirmed due to missing
notification from the remote device. One missing response can delay other
operations by 3 seconds until the watchdog is triggered. This mostly happens
because we either mute or unmute the already muted or unmuted device or set
volume level to the device which already has it set properly. Since the GATT
characteristic does not change, we do not get any notifications, however we
expect them to arrive as an operation confirmation.

The simplest fix for that is not send requests to change volume or
mute state if the volume or mute state is aready set as expected.

Bug: 273276001
Test: atest --host bluetooth_vc_test --no-bazel-mode
Change-Id: Ia60c72552911fdee5b06766fd37154d3ab689d38