leaudio: Yet another reconnection improvement
We want to have non connected set members in the allow list when group
is streaming.
On the event that one device got disconnected we don't know if both
devices are going to be disconnected or it is just single earbud put
back to the charing case.
With this patch, when earbud got disconnected, we schedule a function to
check after 500ms if group is still streaming - if so device goes to
allow list.

Bug: 275753723
Test: manual (streaming, put one bud to the charging case, wait for
screen off, take it back)
Test: atest BluetoothInstrumentationTests
Tag: #feature

Change-Id: I26dc62ec75c8642fe9b5c686563cc0dec57adb38