[Connection Manager] Implement direct + background connections
The architecture is as follows. The LeAcceptlistManager is
responsible for taking a TargetState and driving le_impl to this
target state. The ConnectionAttempts is responsible for
aggregating all the requests from clients, and managing basic
business logic (resolving requests once complete, rejecting
duplicate requests, cancellation etc.). determine_target_state()
converts the aggregated set of requests into a single
TargetState.

The ConnectionManager coordinates the LeAcceptlistManager and
ConnectionAttempts and sends events to them as appropriate, while
offering a top-level API.

Bug: 272572974
Test: unit + manual direct/bg connection from nRF connect

Change-Id: Ibe56093bdccb16246337ce1162ebb43594290309