Remove is_svc_chg flag from GATT discovery
After aosp/2495789, we will only start discovery + read the db hash if
we are doing service discovery. Thus, if the hash read fails, we should
unconditionally do the rest of discovery, rather than early-exiting.

For bonded devices that do not support db hash, we will not read the
hash at all, rather than always failing to read it and then pulling from
cache.

Bug: 273302622
Test: existing avatar tests
Change-Id: I9bad62b6fce46344e0ba2833a6f46ccd2eb64368