btif_sock: Add send_msc interface
RFCOMM layer automatically starts the control request only when it is
the client.

RFCOMM/DEVA-DEVB/RFC/BV-21-C and RFCOMM/DEVA-DEVB/RFC/BV-22-C require
the IUT to be as the RFCOMM server and also the initiator of the flow
control, thus wire RFCOMM_ControlReq to btif to unblock the 2 tests.

Bug: 278505039
Tag: #floss
Test: together with the later CLs
Change-Id: I01f41305458fae839891c13b80286e4296929428