Revert "Avatar: Add ASHA dual device test - test_pairing_dual_de..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/41dfa7dc-066e-4b09-a86c-1d86bfabbd96). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib7003e951e10787326891c4fab709c8887b195c5