Be sure A2DP/HFP profile activated when connected
When a BT device connected that support both A2DP and HFP,
the activation of first connected profile will be delayed until
the other profile connected. If the other profile failed to be
connected, the first connected profile is never activated.
This CL fixes these cases and also activate the first connected
profile earlier based on the audio mode.

Bug: 274741724
Test: atest BluetoothInstrumentationTests:ActiveDeviceManagerTest
Change-Id: Ic462aba67d81e69846f2e829df1b9eae997b54ea