Revert "Adding telemetry for SDP UUID metric choke points"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4248558b-a9e3-4183-afcc-7cd09a27a582). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie4d1d67aeea3f99bb773825f5f51015633c4a5e1