RootCanal: Convert the controller configuration to protobuf
The proto configuration is reduced to simplify tweaking
the controller for one behavior.

The user can now configure multiple controllers served on
different HCI ports with different configurations.

The user can choose a default configuration from a list of
presets and edit the features and quirks to test a particular
behavior.

Test: m && launch_cvd
Bug: 253525123
Bug: 253524429
Change-Id: Ic6fb9aceab9c2246e8aae1f90d51d419d42c5f37