Initialize main thread earlier
A few bug are uploaded and mentionned the DoInThreadDelayed failure
since the log is happening everytime we are booting the stack.
By starting the thread earlier, we ensure it can handle event sooner.

This should be safe to do since the thread is only a tool used by the
stack and is not triggering any event.

Bug: 279365755
Test: Boot bluetooth & check start-up log
Change-Id: I775bee3788559f48da4fb395cdd1bd51cf10619e