Revert "Adds screen on/off event counter"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b47e2b84-4f54-48f2-810e-658e1debee2d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I00c29a0a5f026dc8e027a43ee5216517c2ce0ac7