PandoraServer: Load only Host if Bluetooth is off
If Bluetooth is turned off, the server will try to get the available profiles,
but they will be null. This will cause the server to crash.
To avoid this, only the host should be available. This will allow the user to
perform a factory reset to turn Bluetooth back on.

Bug: 279190568
Test: atest pts-bot:<test> -v
Change-Id: Iaa5c1d5474e58bf5578811edfdf6088ab2d81731