Dual ring-tone played on AAOS
Problem: On android auto builds, both the in-band ringtone & device
ringtone were being played during an incoming call on a bluetooth
connected device.
When the HFP device is in connecting stage, any stack event
for in-band event was being dropped by the state machine. This
resulted in HFP state machine broadcasting information to telecom
framework that in-band ringtone is not supported. Telecom manager
would then proceed to play the local ringtone.

Solution: Defer the in-band stack event until the phone is connected.

Bug: 267975031
Tags: #stability
Test: Tested on device to verify only one ringtone is being played.
Test: Unit-test to ensure message is deferred.
Change-Id: Idaa3b75d82e488c4b3eee96c5e04fe10089cd6f8