BTM: Let CSIS active scanning take effect
Currently when there is an ongoing background scanning like 10% or 25%,
CSIS active scannning would not take effect since the first scanning
come first. This makes finding coordinate members more difficult as the
scanning % is not enough.

This Patch checks if current scan is for CSIS then stops any ongoing
scanning and uses low latency scanning for finding coordinate members.
This increases paring speed and also decreases the failure that we
do not find coordinate members.

Bug: 273958309
Test: make sure we find csis member with 100% scanning
Change-Id: I4b469166690f20306fcdcb57d7ca4a311d5969bc