Revert "RootCanal: Convert the controller configuration to protobuf"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cae7d454-5138-4524-8320-89ce533a0aea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61c7306e2b5da121201453bb57226b5038af86aa