bluetooth: send bonded broadcast when UUID updated message is received on PendingCommandState
When the first device is already bonded and is in the process of SDP, and the second device comes to bonding state, the state machine is on the PendingCommandState.
In this case, PendingCommandState also need to send a bonded broadcast for the SDP result of the first bonded device.

Tag: #bugfix
Test: manual

Signed-off-by: dongguanjun1 <dongguanjun1@xiaomi.com>
Change-Id: I3da3f84eba01fe18cb630580e78d173bf2998ed0