Revert "le_audio: Sync on main loop on MicrophoneAttachToCurrent..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e5802583-fd2d-429a-9c1a-b48dfadbdf2a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaad6568ae678dd46472e1c0964012e348676c075