Always notify on node content, even if something goes wrong
Problem: Some GetFolderItems requests can fail, timeout, or abort, and
we don't always notify the MediaBrowsers of the fact we're as done as
we're going to be with their request. This causes the UI in some clients
to indefinitely assume our service is downloading contents because we've
detached and never notified of contents.

Solution: GetFolderItems is _meant_ to represent exactly one fetch
cycle, so we can move the code more in line with that. We'll always
set the node to cached on exist of the state, and notify clients of the
final contents. In this way, clients will always get _something_,
whether its an empty list or the contents we have so far.

Tag: #stability
Bug: 279516561
Test: atest BluetoothInstrumentationTests

Please enter the commit message for your changes. Lines starting

Change-Id: I6793c61ca91c0ca04aea7b9fc168cf641dee05f3