[GATT Server] Split up Arbiter logic and FFI
The IsolationManager logic is separate from the FFI forwarding, so put
it in a separate module. The idea is that the IsolationManager will
ultimately live in the GATT server module directly, and peers will see
either the global GATT db, or a server selected by the IsolationManager.
FFI will always send connections straight to the GATT module, even if
they are not isolated.

Test: unit
Bug: 274945531
Change-Id: I470d40e7a7c093ff55994aeb6a6935aa854c6970