Allow recovery from an erroneous download attempt
Problem: Failed fetches currently mark the node as cached, which won't
allow another download attempt until the node falls out of scope. This
means the user has to effectively cause the download of another VFS node
that's not a child of the erroneous node and then re-visit the erroneous
node before a new download will be triggered.

Solution: Introduce an erroneous state, which is used to determine if
the cached state of a node was the result of something less than
perfect. When a request is made of a node in this state (i.e. cached by
erroneous), we will uncache the node after its final results have been
delivered, such that a new request will allow another attempt at the
download.

Tag: #stability
Bug: 279516561
Test: atest BluetoothInstrumentationTests
Change-Id: I5623d6b5e3263c2d4a72d1c8d661f6020cdfe81a