Revert "Floss: Caches the discoverable mode"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1fdae32b-18c7-4840-891e-2a36ce414adb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id358bde098cfef799b2d66a7f38a7bad503f34b3