Fix IllegalArgumentException in BluetoothOppServiceTest
trimDatabase() is called in "trimDatabase" thread that's started in
BluetoothOppService#start(). This thread can be ran even when provider
is not enabled, and even not in the same time frame with the current
test. This might raise exception, resulting in the tests following the
test that initiate this thread to fail.

This cl also fix similar problems with other threads like mUpdateThread
and NotificationUpdateThread

This cl also fix memory issue mentioned in b/266459454, where
ContentProviderClient is used but not freed

Also add more tests for BluetoothOppService

Test: atest BluetoothOppServiceTest --rerun-until-failure=100
Bug: 281311423
Bug: 266459454
Tag: #refactor
Change-Id: Ieb8129bf29a7e1841ba3d6c40f677cec7dd9b897