Revert "Unit test for btif avrcp audio track"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/56c91cc9-02f0-4d03-984b-17388f36c568). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iec79272c07e92c681ee22c7aa526a9b1c746008c