Revert "Unit test for btif avrcp audio track"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8b756633-d00c-40ac-9ed9-2c32fd8089df). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib6495ec6908194ffe301485dd3c803e13736a3ec