Revert "Unit test for btif avrcp audio track"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4b48e8de-578b-46c3-8996-97990e0e4bd6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3004abc0976285fb0a40e75cb889289a4ea9d1bc