Revert "Unit test for btif avrcp audio track"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d7cd5adf-a97a-4872-bf8b-6f54ad475702). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7dbb48b9fabd3a7ea48b8b5880a0fd56b211bf6c