Revert "Unit test for btif avrcp audio track"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/63969e5c-6de1-405e-82a4-c82eac81dc76). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I39ef4917acb058eefdbe7f23731502d7dd7039aa