Revert "Unit test for btif avrcp audio track"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7e3f949e-5b10-4770-9f21-346a11e3ecbd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7eef6048e751e68e690af784c8fef9ce9861ae5f