Revert "Unit test for btif avrcp audio track"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c46b392-14ba-451b-8945-21a5e1ee3b35). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifd3315d0019686846ba2782417c1ef59d100ad56