Revert "Unit test for btif avrcp audio track"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f611339b-d278-45c4-97f2-cd807a909408). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie82d949a34e1b5f51a3dc3642b239aeb22f66a4c