Stop using BT icon from private resource
Currently the BT app icon is a private resource in frameworks/base.
As a result, the icon is not properly shown if the resource ID becomes
different between when mainline module is built and when it is run.

This CL copies BT app icon into BT module and uses it.

As the tint color in bt_share.xml is also private,
this CL uses a blue color from official Bluetooth website.

Bug: 279384742
Test: Sharing via bluetooth shows the correct icon
Change-Id: I9d2ac626f294dea17eecebfd70cd1e7a2ae01cd0