Add direct Telecom check for call state in HFP service
Telecom is using BluetoothHeadset#connectAudio to start
SCO, but HeadsetService checks if a call is ongoing (or
ringing) before sending the command to the native stack.

It can happen that BluetoothInCallService isn't yet
connected to Telecom service when that connectAudio
is called, resulting in SCO not started at that time
but when BluetoothInCallService is connected instead, which
can be an issue as we will return false at Telecom's
connect audio request.

Bug: 274081739
Tag: #feature
Test: manual
Change-Id: I74b8f56912cbd03e68321db1c5389e1374bca9bd