floss: Connected devices skip acceptlist in bgconn
When a gatt background connection is added, check to see if it's already
connected and skip the acceptlist if it is. When the device disconnects,
if it is in the background connect list, add it into the acceptlist.
This prevents devices with RPAs from getting connected twice (once with
the random address and once with the identity address).

Bug: 271827442
Tag: #floss
Test: CTSV BLE Encrypted Client Test
Change-Id: Ief8463f3128fe81702d3cd5fe8ec52760e3a11ea