Revert "pdl: Generate snapshots using prettyplease"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1008c06a-ea41-4ec2-a49a-26523d886cbf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4bb90f25b16cc39385489c94d0cc36f9df96eeb7