Revert "pdl: Format generated code using prettyplease"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5902032e-23e7-42b6-b3c2-6cc9b7c537df). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If16f692dd98d8ace64a2770bd52a95b84183e71a