Revert "pdl: Generate snapshots using prettyplease"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5902032e-23e7-42b6-b3c2-6cc9b7c537df). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6c113bd2b83bbfa9f1960efdbbf75e43bd527497