Revert "pdl: Generate snapshots using prettyplease"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/18719de8-330b-49c3-a3be-fa28a269d7e4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0761e08c3244e5adf17626223a436c0ffee38286