Revert "pdl: Format generated code using prettyplease"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c0bc3733-e3d8-4c3c-9e69-ef27b2b7f819). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2d58c3a8aa4bb05d4dcc0f7b797f328686dcf205