Revert "pdl: Generate canonical tests with prettyplease"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c0bc3733-e3d8-4c3c-9e69-ef27b2b7f819). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id478ea7906b570893df1b6ecc222c7787d6f6c54