Revert "pdl: Remove rustfmt from no_alloc test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c0bc3733-e3d8-4c3c-9e69-ef27b2b7f819). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7171fa85cd663fc55893aa0b5bf37366e4ee1ddc