Revert "pdl: Generate snapshots using prettyplease"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e898b117-7275-42f3-b669-d81bee33fb11). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If46f8003675a18eaeccd5af4756e5a55e6508dff