Revert "pdl: Remove rustfmt from no_alloc test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/887bb3c3-1b68-4747-9c27-ebcabc6d0ccc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5eb9a916faa0449dbc22851a983bdbaa78797ead