Revert "pdl: Format generated code using prettyplease"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/887bb3c3-1b68-4747-9c27-ebcabc6d0ccc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6bd97e89dbb8c4a1c3f589410732d06b4a11136e