Revert "pdl: Remove rustfmt from no_alloc test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/866286b5-38d8-45ce-b1de-99984709c4b9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I99c8e6cc7109969d2833b8442e2401cd0fbc392d