Revert "pdl: Remove rustfmt from no_alloc test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e8ca4882-2bdc-45fc-925f-76df8dfebf3e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic4c1e38fd164c21e48d801be5b1c87b3d6448aae