Revert "pdl: Generate snapshots using prettyplease"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e8ca4882-2bdc-45fc-925f-76df8dfebf3e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0cd84d1c54dd14361c092ae730b8500468183e6d