Revert "pdl: Generate snapshots using prettyplease"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6bbe77bf-ac28-4a33-b970-e98e68e3ca08). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7a64a3db6bd682ffcc063b7081b774142d86e631