Revert "pdl: Generate canonical tests with prettyplease"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6bbe77bf-ac28-4a33-b970-e98e68e3ca08). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5c284a8ea845942b3e92137773611a7af2e7d72b