Revert "pdl: Generate canonical tests with prettyplease"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd7b66d4-3125-40c2-8bdc-783f6d4b1ec4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I543415cb3e0709e9c700bcbd75b9a7adf12241bc