Revert "pdl: Format generated code using prettyplease"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f35f6679-b533-4c38-878a-137818f3176c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4f667c431e8f860cbcfd607c7abf953b7c5ea4f1