Revert "pdl: Generate canonical tests with prettyplease"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f35f6679-b533-4c38-878a-137818f3176c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1ffb1e5c3516063dae55dbc1807c0ebcf7f2965b