Revert "pdl: Remove rustfmt from no_alloc test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a8db814-7c8a-494f-be7a-0c7b878117ca). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iefb932ceb104f893a56701e701cdc31546ba2748