Revert "pdl: Remove rustfmt from no_alloc test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/754bd395-37e1-4db4-87a4-03969700da83). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id251011aafd847ee2fb3c558ccd12618d8cee9df