Revert "pdl: Format generated code using prettyplease"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2439bd55-e8a3-4a16-91e4-08e54f0d6e46). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia89f26df3b11e11c280cd1389f588bd9b91e6392