Revert "pdl: Remove rustfmt from no_alloc test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2439bd55-e8a3-4a16-91e4-08e54f0d6e46). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I622fd10d8851c484304ab0a1d3205132c4127b48