Revert "avatar: update after offical tests move"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0b241e6e-b6cc-480f-b30c-de624f41aa72). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61c0cbcd580cd17c5647f6755f36a3e6899a2a36