Revert "avatar: update after offical tests move"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cc1039d9-0272-47a9-b3a4-d10f90de2963). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iffafcc8ac3c00cba465dba7a7a825b0842ed6337