Revert "avatar: update after offical tests move"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fb7da6d3-7917-49d2-aff3-0742184420e6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibc5a1fa2d1828516bf71995d41f41256a7a72561