Revert "avatar: update after offical tests move"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1b6a0afc-7f13-4e6d-9ebd-8a49c20bf69a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I146bb25b35c253312d129662e375afac41caa4ad