Revert "avatar: update after offical tests move"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c397013f-d1ca-4b98-ad75-bf609cbf333f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie3f1f85151984ade5982bbbccfff197754af1866