Revert "avatar: update after offical tests move"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/84dcb976-41bd-4def-b82a-2a4aa75a5a9f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5dc0c0a69630af53bd592114bae6fbb3f16cb301