Revert "RootCanal: Tie the random number generator to the dual m..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6a3344e2-7d2f-4740-a87e-b158aba53bed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iec440465663762d8d82e1e88217a80296c849763