Revert "RootCanal: Tie the random number generator to the dual m..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/10dce38d-1497-4a8d-bc82-86dac15b17f0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I58bd9865f14d34e6a34f53f76d627dc9268b734c