Extract AdapterBinder from ManagerService
This allow to catch external call and simulate the AdapterService
without the need to rely on an the real implementation of bluetooth.
There is no need for native call / hal setup and this reduce the
potential falkyness of test in the Manager Service

As a result this CL contain the first BluetoothManagerService test that
simulate a start of the bluetooth until it reach bleOn or On.

Bug: 262605980
Test: atest ServiceBluetoothTests
Change-Id: Id528d28c0a623f5fbec744bbc79fcca822602d22