Revert "rust: Update to syn-2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/36c40f0b-3512-4d33-b6fc-40b4bdca3539). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic424192ce1e7ca81ff79e4702d51e856fb6b179a