Revert "rust: Update to syn-2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/50c968f8-8ab0-44bc-a256-408e27d0a777). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0f561e97ab62f00b16dcdd9a940ff45f5613ef23