Revert "rust: Update to syn-2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/50c968f8-8ab0-44bc-a256-408e27d0a777). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib89abe2427f1b01efb74e45e4f4af5fc8eaa012a