Change callback name to more meaningfull name
Previously both these method were used to switch from Ble to BrEdr or to
off. The name `onLeServiceUp` is not accurate as the AdapterService is
already sending a "le is up" notification to BluetoothManagerService
with `onBluetoothStateChange(x, BLE_ON)`
Changing the name to not be a callback but the acutal action asked to
adapter AKA startBrEdr.
Same for onBrEdrDown that will instead be called stopBle

This is a no-op change

Bug: 262605980
Test: None
Change-Id: I7d55925e6a34f7f0a9eb2d0cf338ad5cc74d678f