Change callback name to more meaningfull name
Previously both these method were used to switch from Ble to BrEdr or to
off. The name `onLeServiceUp` does not mean anynthing and it is
preferable to use a name that explain what it does AKA startBrEdr.
Same for onBrEdrDown that will instead be called stopBle

This is a no-op change

Bug: 262605980
Test: None
Change-Id: I7d55925e6a34f7f0a9eb2d0cf338ad5cc74d678f