Revert "RootCanal: Remove h4_packetizer_lib to device/google/cut..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/02b00979-5fa1-45dc-b59a-43455b87250e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2d43a18d2f12a631fe641d365a0f67d551d278e0