le_audio: Split AudioStreamDataPath to Data Path and CIS state
There is possibility to have more different states than covered by
AudioStreamDataPath. CIS state is not strictly related to DATA PATH
state. There is also a possiblity to have disconnected CIS and
configured data path.

as per Bluetooth Core v5.4, 7.7.5 Disconnection Complete event,

"If Connection_Handle identifies a CIS on the Central, the handle
and the associated data paths of that CIS shall remain valid
(irrespective of whether the disconnection was successful or not)"

Bug: 278621595
Tag: #feature
Test: atest bluetooth_le_audio_test
Change-Id: I9b363484ba881ce6f94017a2a72b7bb6cb826d9e