Tbs: Fix for possible IllegalArgumentException
When device got authorization for to use TBS it send notifications on
all the characteristics which are notifiable.
However, some characteristics might not have value which would lead to
java.lang.IllegalArgumentException: Characteristic value must not be null

this patch fixes that.
Also, value of Status Flag is kept in the mStatusFlagValue Map and it
should be used when sending notification.

Bug: 285528071
Tag: #feature
Test: atest BluetoothInstrumentationTests
Change-Id: Ia64fe3718448d07b4741b29eac67f5560fb28dab