Read AirplaneMode only once by event
Multiples call to settings.Global can return different value when the
setting is quickly toggled on and off. By reading it once we ensure the
state we want the bluetooth to be in is computed using the correct
value.

Without this fix, we could turn off the bluetooth while listening to
media.

Bug: 271963349
Bug: 262605980
Test: atest ServiceBluetoothTests | added testFastToggle
Change-Id: I46366527eb0fa1cc71f789334c4241241f9c9274