Battery state machine goes into WAITING on timeout
The battery state machine connects to
a remote GATT server in an opportunistic way.
Therefore, there is a chance that the connection
is made after the timeout, which is totally valid.

In this case, the battery state machine goes into
the WAITING state instead of the DISCONNECTED state with the change.
This will allow the battery state machine
to wait indefinitely for the connection to be established.

Bug: 261465867
Test: atest BluetoothInstrumentationTests
Change-Id: I522da76de9666db8384f51f72f6078fd8f8dd60e