Floss: disconnect all profiles when media profiles are cleared
We used to disconnect only media profiles when there is an error during
media connection.

In this CL, we disconnect all enabled profiles when a critical media
profile is disconnected.

This also makes the `disconnect` function call-able as an entry point
because while it may not disconnect all media profiles at once, it
will disconnect all profiles in the end after some iterations.

Bug: 285391400, 286330209
Tag: #floss
Test: cq

Change-Id: I2edb9f74a7b71800178336b6dae80a80d3e5db5a