Revert "Remove unused stack::btm_api_types::BTM_INVALID_SCO_DISC..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/75f7e96e-64a4-43c7-8ad4-6b9d9e11f6e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6e92c9241065e820bc2c214325958af40990caad