Revert "Remove unused stack::btm_api_types::BTM_INVALID_SCO_DISC..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2d0624d7-79bf-4e26-a3ee-6fe7a44b13c7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib77976365884ff74a45333b70faa695e01416d98