Move shared_library with a static variant to static_library
APEXes embbed copies of transitive dependencies meaning that
most shared_libraries are not shared with the rest of the system.
In this case static linking is a net worth as the dependency is
still copied but we allow the linker to do more optimisations.

This reduces the APEX size by 1.6M.

Bug: 265285999
Test: m com.android.btservices
Change-Id: Ifb6dc5511a6d22b14d0a54e100f12aec43c5ae3b