Revert "Properly type in bta::tBTA_DM_SP_CFM_REQ::rmt_io_caps"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8e4d6b03-a5ba-4735-9387-85a02e941f69). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia59cd559a78fb6fc99792441ec70d0809f296e69