floss: Exit start_scan early if already scanning
On a Dell Latitude E6230 with a Bluetooth adapter BCM20702A0
(vid:pid of 413c:8197), pairing with a Logitech MX Keys Mini keyboard
fails often due to receiving a 'Command Disallowed' response to a
Set Scan Enable request. This then causes the `ASSERT`s in
LeScanningManager::impl::check_status to fail, resulting in a crash.

For unknown reasons, this failure is avoided if we refrain from
sending another "LE Set Scan Enable" command if we are already scanning.
This early check is similar to the early check currently performed in
`stop_scan()`.

Bug: 287278110
Tag: #floss
Test: Deployed this change on ChromeOS Flex to a Dell Latitude E6230
      containing the Bluetooth adapter 413c:8197. Enabled Floss and
      successfully paired to a Logitech MX Keys Mini keyboard on
      every attempt (at least 10 attempts were made).
Change-Id: I85d5a57d3ae53135b9bbe4e40ee0dac2cae2fc15