LeAudio: Decouple configration mechanism from the available contexts
Current audio set configuration should be coupled with the current
audio use case, therefor coupled with the current call state, sink and
source metadata if any of these are activated. It should however
not be coupled with the remote available context as even when the
remote does not declares for example Live context availability, we
still want to support that scenario but under a different audio
context that the remote device supports.

The existing configuration logic is coupled with the context
availability at many levels and has to be reworked.

Bug: 285647765
Test: atest bluetooth_le_audio_test bluetooth_le_audio_client_test
Change-Id: Id2821ef1681e99f3c5afe109473c43abf91de670