Bluetooth: Sleep before crashing when the HAL dies
The HAL crashing is an unrecoverable failure. Sometimes, the system kills
the HAL as part of a system shutdown, in which case Bluetooth will also
be killed. Sleeping for one second after being notified that the HAL
died is a way to make sure we don't report deaths that are a result
of the system killing the HAL out from under the Bluetooth stack.

Bug: 284909883
Test: mma -j32
Test: Kill the hal, then Bluetooth with different timings
Change-Id: Ie5b3c2c9f47a11896812b1f07a7da664cc84f3db