Revert "SystemService: Promote more warning as error"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a524d07e-db6f-4e3a-9b6a-ec6de89fde92). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If00653b966c87169cf74ae1898b9fd613e158ca3