fix BLE scan could be started even screen is off
AOSP logic only allow APP to start BLE scan when screen on if
APP doesn't set any filters.

But if an APP set a filter with all fields empty, AOSP logic still
think it's a filtered scan, it's a bug.

The logic should check whether client's filters are empty and
whether each filter's fields are empty.

Bug: 290856528
Test: atest testScreenOffStartEmptyFilterScan
Change-Id: Ia9471cb7cfea684cd90ddd4626dc3c3932793af8
Signed-off-by: yinzl1 <yinzl1@lenovo.com>