IA2: Add device address to stack::sdp callback
The current codebase presumes that only a single
SDP cycle may be in progress at a time.  When initiating
multiple SDP requests in parallel the bd_addr will
be necessary to align the callbacks with the outstanding
SDP requests.

This CL simply provides a currently unused bd_addr
in addition to the SDP status upon completion of
an SDP cycle and possibly user data in those cases.
As of yet the us this additional parameter is unused
but will be required for parallel SDP cycles.

Bug: 291117499
Test: mma .

Change-Id: Ib503b518be870013a8728410072e9708a9385849