Check the validity of hiSyncId when adding device to the set of aid hearig active devices
Unpair the aid hearing device, set aid hearing device to null, so add null to the set of
aid hearing active devices without checking the validity for hidSyncId.
When a new a2dp/hfp device is connected, the set of aid hearing active devices is not empty,
so donot set a2dp/hfp  to active.

Bug:
Test: manual
Change-Id: I4bf6f2e8acb0f8ffa2efb8ac9d247617988d3f67