LeAudio: Change the initial configuration
Let's start with the UNSPECIFIED configuration, which is technically
the same as MEDIA, but allows otherwise not reconfiguring scenarios
to actually properly reconfigure for that use case, instead of assumming
some scenario without the knowledge of the actual past scenarios.

Additionally this will allow to properly unit-test the scearios which
would not reconfigure from MEDIA due to our reconfiguration-minimizing
policy. Once the initial configuration is applied, this policy will
be applied as usual.

Bug: 285647765
Test: atest bluetooth_le_audio_test bluetooth_le_audio_client_test
Change-Id: Iaccaa3b70233a518071c7a82e22b26e827d17f27