LeAudio: Workaround VoIP calls not using Telecom API
When the Telecom API is not used for the VoIP call, we do not
receive the InCall mode callback. We need to figure out that this
is a VoIP call and reconfigure for the CONVERSATIONAL context. However
since Telecom is not involved, TBS does not know about the call and it
will not be able to control the call state. We should also NOT put the
TBC content control ID into the stream metadata, as there will be no
incoming call on the TBS service. Otherwise the remote device would
get confused.

Bug: 284230664
Test: atest bluetooth_le_audio_client_test
Change-Id: I4d320e71055a06657c9dd688a6f873916cf862ec