floss: Fix hal host startup race
When hal host starts, it calls 'Read Index List' which will return all
the available indices. Depending on timing, it's possible for a 'Index
Added' to be emitted before the 'Read Index List' command completes. In
this case, we should simply continue (and wait for 'Read Index List')
rather than exit with an error.

Bug: 271899772
Test: mma -j
Test: emerge-brya floss and start/stop stress test on Chromebook
Change-Id: Ice5d13c7b3e5f3a76e97144ae09e4463efec0f51