To check Remote device address added for Interoperability feature.
Some remote devices are misbehaving when there is an active
call and MT call is in progress. We send call indicators for
active call and MT Call in progress.
When active call is terminated,
we send call end indicator only to remote.
This is confusing remote and remotes are going into bad state.
Denylist the remote devices to disconnect SCO when active call is
ended, and send fake MT call indicator again with some delay.

Change-Id: Ie36fe5e368df7bb25b239d10061b5428369e3a20