To check Remote device address added for Interoperability feature INTEROP_HFP_FAKE_INCOMING_CALL_INDICATOR
when there is an active call and also Incoming call in progress
Host sends call indicators to the remote for
the active call and the incoming Call in progress.
When the active call is terminated,
host sends only the call end indicator to the remote.
This is confusing few specific remote devices
and the remotes are going into bad state.
Adding these remote device address in the interop database file
and checking the device name from application layer to disconnect SCO
when active call is ended, and
send fake incoming call indicator again with some delay.

Change-Id: Ie36fe5e368df7bb25b239d10061b5428369e3a20