Remove role condition when binding BCB to CCB
Problem:
AVRCP/AVCTP connections typically connect control, then browse.
AVRCP uses the AVCTP API to create an FF:FF:FF:FF:FF:FF (any)
CCB to allow incoming connections, listed with role=ACP. When
it makes its outgoing connection, it makes a _second_ CCB with
the specific BD_ADDR it wants to connect to and role=INT.

When a collision occurs, the outgoing connection has an LCB made
and associated with the INT CCB. The incoming connection then
gets processed and has it's channel ID written over the
_outgoing connection's_ LCB, and the outgoing connection's
channel is moved to a "conflict" spot in the struct. This has
the side effect that the outgoing connection's CCB _and role
information_ are now associated with the incoming connection,
which makes the future check when browse is connected ultimately
fail permanently because it wants role=ACP (which we were).

Solution:
A quick and dirty solution is to ignore the role and bind the
new BCB to the first free CCB we find where there's a control
connection and a null BCB pointer. This is effectively saying
that we don't care where the connection came from, just that
there ends up a control and browse connection with the device

Tag: #stability
Bug: 285621447
Test: atest BluetoothInstrumentationTests
Test: 100 iterations of connect/disconnect -> 100/100 success
Change-Id: Ia2216d4b1261f78c4bee1b5599743f9f8509ec82