Floss: Hcidoc: Attempt to identify Nocp for unknown handle
Currently when a device is disconnected and the handle is unknown
(happens when the connect complete event occurs before the start of
the snoop log), we don't report the potential Nocp. However, we stored
the ACL transaction for every handle regardless of whether its known
or otherwise, so we have enough information to actually determine
whether the disconnection has Nocp issue.

This CL doesn't skip checking Nocp for unknown handle.

Bug: None
Test: Test with report 91002174092
Change-Id: If4b768c56faaf22845d69fe044606eb830a403ba