Update playback state when track changes
The active queue item id is set when the track comes in and is set on
player object internally, which hands over playbackstatecompat objects
for our service. However, we don't update our service right when we get
the ID. We wait until the next playback state change happens to update.
This is typically invisible, but updating it sooner is the right thing
to do.

Tag: #stability
Bug: 293511407
Test: atest BluetoothInstrumentationTests
Change-Id: I5ea7b35511e61d2817b6017fbefb6a3856bef217