activityAttribution: Remove dead code
Since initJni is never called, the callback are never registered and
therefore are not usefull. Everything related to the callback can be
delete
The cleanup is a no-op call, and can be removed

At the end it only remmain the notifyActivityAttributionInfo.
We need to check if the feature is working as expected or if it is
deprecated

Bug: 291187783
Test: m Bluetooth
Test: atest ActivityAttributionServiceTest
Change-Id: Iaa590539c63623e5f22e23df4b0e177ce45240cc