Add random delay for AVRC connection request
AVRCP Browsing connection collisions causes a disconnection
from each side. As Android Auto has the same delay as
Android phones, there almost always is a connection
collision.

There is a current issue where after the collision, we
don't accept any incoming connection requests as we are in
initiator role, so to unlock connections in the meantime
we are adding the delay by default.

Once the collisions are handled properly, the delay will
happen only on the retries.

Added a mock for osi_rand needed by the tests.

Bug: 285621447
Tag: #feature
Test: m libbt-bta
Test: Manual connect/disconnect with AA device
Test: atest bt_host_test_bta
Change-Id: If3b15a430bd6a7c94eb79886e7646232eaf4155f