le_audio: Protect from unreliable autonomous disable operation
Both ASEs that are bonded with bi-directional ASEs should go to QoS
state once CIS is terminated. This change will guard for such
transition.

"If the server detects link loss of a CIS for an ASE in the Streaming
state or the Disabling state, the server shall immediately transition
that ASE to the QoS Configured state." ASCS 1.0 3.2
ASE state machine transitions.

Bug: 278621595
Tag: #feature
Test: atest bluetooth_le_audio_test
Change-Id: Ic6e91d63b9690771bd7da09f1a108a181369b861