Don't add duplicate elements in LRU queue and keep synchronized with mDevices
If you pair and unpair the device multiple times, the device will be removed from Hashmap mDevices when unpairing. The counting queue mDeviceQueue is not removed synchronously. When the number of times is large enough and the storage in mDeviceQueue is greater than 200, the first element of the queue will be deleted. Since the same device was saved in the queue, it was removed soon after joining, resulting in the failure to obtain props. So there is a same element in the ArrayDeque, no longer add repeatedly.

Bug: 294781431
Bug: 305330174
Test: manual
Change-Id: If7ed2900bf7f12edf83c2b1760bc665fd06ea39c