Revert "RootCanal: Use unique device identifiers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ba8249de-477d-41d8-8f5a-8c9616630e99). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If61e9b9166b142f35280f0e59ace2381072a56ec