Revert "Stop using intent from LE audio profile to ActiveDeviceM..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/86731869-c5ef-4089-aa98-c06346ff867a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I043c55c2af87ed09b9ab9142fc0968df5e269eaf