Floss: Add Disconnect method to bt_sock interface
When disconnecting a peer device, we first disconnect all of the
connected profiles. However, if the profile connection is created via
a listening/connecting socket, floss layer doesn't know the current
state of the connection, since the corresponding file descriptor has
been passed to whatever client that asks for the socket.

Therefore, add a disconnect method to the socket interface to ask
libbluetooth to disconnect all socket connections which corresponds
to a particular device address.

Bug: 288939390
Tag: #Floss
Test: Fast pair Pixel buds pro, disconnect from peer, then press
      disconnect from UI about 2-3 secs later. Verify the device is
      disconnected.
Change-Id: I5927fdd4d2a99240af5f007564b0f5b42136c869