Revert "Move BTM_VOICE_SETTING from header to source"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b11b3548-82ea-48a3-a367-f5d79bbe51c2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5bc592dd3416f696edcb2145ac1ac5396155d295