Revert "[le audio] Broadcast use non-resolvable address as default"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/10e1595a-6180-458b-aef8-155420ad21ec). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id3ed41d0e50007f6829f9d4eec502e020b8bc06c