Revert "Check for valid ACL disconnect reason"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/48992a10-3850-4c09-888b-a11d94622e93). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia1a39d33e91e2cd603023fe83e121e7b989eca36