Revert "Check for valid ACL disconnect reason"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/edcf5e96-d3bc-4e3d-86ee-9a94f09a9873). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1e54d5676465b86e1c3564c46f60c42f8346b5a2