Revert "Add waiting timeout in test to fix flakiness"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/07ca096b-f5e2-40c8-9614-f09b47fb8b12). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic6707a81689859cd9b521f4e477d8522318fcc89