Clear the status of the direct connection when canceling BG connection
Problem: No longer try to connect to the device after background connection is canceled
Cause: If the background connection is canceled during direct connection timer to the same device, only the direct connection timer is deleted and the callback is not called. As a result, it is recognized as a device that is already connected in gatt layer, and tcb is remained even after that. It is no longer possible to connect to the device.
Measure: Clear the status of the direct connection when canceling BG connection

Test: 
1. Attempt direct connection to device without advertise
2. BG connection before direct connection timer expires
3. Cancel BG connection
4. Direct connection is not possible even after a while

Change-Id: I1721524e30956f8a11dfab4e5a7c3c80e81402b7