floss: hcidoc: Add Authenticated Payload Timeout Expired event
This patch adds the HCI_Authenticated_Payload_Timeout_Expired event,
which is used to indicate that a packet containing a valid MIC on
the Connection_Handle was not received within the authenticatedPayloadTO.
A host may choose to disconnect the link when this occurs.

This event can be used to investigate disconnections that may be
related to authenticated payload timeouts.

Bug: 298641694
Tag: #floss
Test: ./build.py --target (test|utils)
Test: run hcidoc with btsnoop log in b/298641694
Change-Id: Id1d506da6071c7d720b77f7328708874fb7ffe7e