[LE Aduio] Handle later join case for 32k bidirectional dual mic
In conversational mode, when we disable 32k bidirectional dual mic,
we select other configs like 16k bidirectional dual mic.

However, when users connect to only one bud while they put one earbud
back into cases, we select single device on 32k, which is not suitable.

This is because CIG is already configured for 16k but when there is
only one device connected, we choose 32k for single device, and this
will make us removing CIG and create CIG again.

The user experience is not good as the audio will stop for some time.
This patch handles this case.

Also, not insert config to the mpa when it is empty.

Bug: 298946279
Test: check early exit and later join on conversational streaming
Test: atest -c bluetooth_le_audio_test bluetooth_le_audio_client_test
Change-Id: I50a0ccc0906a2c3edefcfa553224498e87a54267