Floss: Hcidoc: Relax the no NOCP detection on baseband disconnection
The controller might not have the time to send NOCP packets for an
outstanding ACL packet if there is an abrupt baseband disconnection.
This disconnection is not caused by NOCP issue so we shouldn't mark
it as such.

This CL doesn't flag such a disconnection as NOCP issue if the
disconnect event comes less than 1s since the ACL is sent.

Bug: 293397988
Tag: #floss
Test: Verify feedback 91179429567 is not tagged NOCP at the time
      12:04:26.598551
Change-Id: I84eaa3da1fbb659f70ab22f834ce2843727a3f89