Floss: Ignore multiple RemoteHostSupportedFeatureNotifications
It is observed that the host controller can receive multiple Remote Host
Supported Feature Notification events during Remote Name Request stage.
This causes a crash since the callback function is BindOnce. Although we
are not sure if this is a controller error, peer devcie's fault, or IOP
issue, it should be harmeless to just ignore the second feature
notification and proceed the RNR.

Bug: 296988044
Test: mma -j32
Test: intentionally to invoke
on_remote_host_supported_features_notification twice and make sure there
is no crash.
Tag: #floss

Change-Id: I17770b49e60e719fbeedae81eac88dc5c6b45ab1