Allow encryption request to proceed in unrelated states
Device security state is shared for both transports. Encryption request is saved if security state is not idle.
This means that security state for a transport can block encryption request the encryption request for another transport. Saved encryption requests are revived for just updating the requester of encryption change status when encryption change is event is generated. Thus the saved encryption requests are never acted upon.
For, example if LE encryption is requested while the BR/EDR link is disconnecting, the LE encryption request is blocked and never acted upon.

Change-Id: I7d29f6575d0b5ed3578051d800976c5f3149e552
Test: mma
Bug: 294886212