Revive the enqueued encryption request
Encryption request are queued if security state is not IDLE. Security
state is shared for both transports. Saved encryption requests are
popped out when encryption change event is received for the same
transport. However if the encryption request was saved in a state which
does not generate the encryption change event, the encryption request is
never revived.

Test: mma
Bug: 294886212
Change-Id: I1ec17d0ad06e752494a11bb51d3c746713ee6cd9