LeAudio: Rename codec specific structure
To not confuse CodecCapabilitySettings with the PAC service
codec cpabilities, and since this struct is being used to pass
the codec specific configuration (in codec manager and audio
set configuration provider) it was renamed to CodecSpecificParameters.

In future this should wrap not only tha base configuration for the
already known codec specific parameter types, but also the vendor
specific ones, thus will migrate from LeAudioCodecConfigBase to more
generic LTV map.

Bug: 295972694
Test: atest --host bluetooth_le_audio_test bluetooth_le_audio_client_test bluetooth_test_broadcaster bluetooth_test_broadcaster_state_machine bluetooth_le_audio_codec_manager_test
Change-Id: I89faef9f2facb7a498f972ffe1562ffd87a09d5f