Revert "Remove unused headers from hci/include"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/56b0c025-33ea-4468-88e1-f360d7242332). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3d4be238e790fc416750dd5e4782ce33c4eb6b2e