Revert "Remove unused headers from hci/include"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f20580c2-b3f5-4cb3-aa75-a8476c0ffbc5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iac2ef951b34b65c03520635acf0104c80b63fea8