[le audio] BassClient enhancement to allow re-sync with known inactive source
In current design, assistant will not be able to add source for inactive source.
This commit implemented a solution to allow re-sync to previously cached source.

When syncing with a new source, the assistant will cache the scan result. Then if later
the device stoped the sync or user chose to remove the source, adding source operation
will check the previous cached scan results. If the source has been found, the assistant
will try to resync to this known remote source.

Bug: 299422016
Test: atest BassClientStateMachineTest
Test: manual test with broadcast
Change-Id: I09f0c261fd5c937fca75f84046199d6aa6d2c768